Fix error on empty CSV with header and transforms
This was specifically an issue when the transforms dictionary included
functions that didn't have known return values for String inputs and
when there was a header row but no data (or when the buffer was empty
but a header was specified). DataStreams.transform attempted to infer
a return type for the function and got back Union{}, which errored.

Closes #74