Fix and improve CategoricalArray handling (#293)
* Fix handling of CategoricalArray levels

The old code called push_level!, but that internal function does not add the level to invindex,
giving a corrupt pool. Use the get and get! public API instead. This involves a second lookup
when inserting a level, but the performance impact should be negligible.

Also ensure levels are sorted each time we add a new levels, so that the relative order of
existing levels (which are exposed to users when streaming) never changes. This operation
could be made more efficient by storing the information in CategoricalPool, but that would
make a difference only when the number of levels is large compared with the number of values,
which is not the use case CategoricalArray is optimized for.

Test added in the next commit.

* Fix passing CatValue in types arguments

Without this we get an error trying to access the pool.

* Removing percentage of unique levels logic when categorical=true

Instead, always return a CategoricalArray for simplicity and reliability.
Also change the type checking logic to work around a type inference bug.