Use pool=0.1 by default
CSV files frequently contain variables with a small number of levels, for which
common operations like grouping and element-wise functions are much faster if
data is pooled. Pooling requires hashing values during parsing, but that cost
is worth it since if this kind of operation is performed at least once.