Fix #578: multithreaded parsing on really wide files
The issue here is that when we chunked up the file to determine the
starting position that each thread can parse, we forgot to account for
any skipped rows, which led to the unfortunate scenario in #578 where
the 1st thread technically started on the correct row 3 of the data, but
the 2nd thread was actually starting parsing at a byte position _before_
row 3.

Another issue I noticed while debugging was the "file" was estimated to
have 12 rows, which is pretty accurate, but then when we chunked up the
file for multithreaded parsing, each thread only expected 1 row
(essentially it did `div(12, 8)`), which led to two of the threads
having to re-allocate while parsing to fit the 2 rows unaccounted for.
By switching the calculating to `cld(12, 8)`, each thread plans on 2
rows and we avoid any re-allocating. For very wide files like in #578,
this is a tad inefficient because each extra row is really wide, but in
the general case, this should avoid re-allocating a little more often.