Fix multithreaded column type promotions
The issue here is after we've done multithreaded parsing over chunks of
the file, we merge each threads' column types detected to get the
"final column types". The bug here was we didn't have a specific rule if
an earlier thread had detected `Union{Int, Missing}` and a later thread
detected `Float64` or `Union{Float64, Missing}` (and similarly for the
reverse situation, float detected earlier and integers later).

cc: @andreasnoack, @pkofod, this should fix the issue in the file you
shared with me privately