Integration column selection directly with parsing (#585)
* Integration column selection directly with parsing

Implements #557.

We've always had the ability to do fairly efficient column selection
post-parsing on a `CSV.File` object via `Tables.select` (now
`TableOperations.select`), but this still meant parsing the full file,
which, depending on the column type, could be a little costly.

With this PR, we introduce the `select` and `drop` keyword arguments to
`CSV.read`/`CSV.File` which allows passing a list of columns as an
`AbstractVector{Int}`, `AbstractVector{Symbol}`,
`AbstractVector{String}` or `AbstractVector{Bool}`, or as a "selector" function of the form `(i,
name) -> keep_or_drop::Bool` to specify which columns to "keep" or "drop" after parsing.
Internally, this utilizes a trick we already had in place wherein if a
column type is user-specified as `Missing`, then parsing efficiently
"skips" any type parsing for that column (basically just moves our
parsing state past the cell). So in terms of implementation, this is
surprisingly simple: for any column we don't want, we set
it's column type upfront to `Missing`, parse the file, then remove any
mention of the dropped columns in the final `CSV.File` object.