Fix a few issues for (USER | MISSING) column types (#594)
Fixes #592. With the new `select`/`drop` keyword arguments, we'll now be
seeing many more column types w/ typecodes like `(USER | MISSING)`.
Which are handled specially when parsing. In particular, the parsing
tape is allocated with length 0. For parsing a column of this type, we
just consume the cell in the file without storing anything. The problem
in the original issue was in the tape re-allocation code, we assumed
every tape had length to copy from old tape to new tape. There was
another issue in the invalid row checking code where if a row didn't
have enough columns, we usually set the value of each unparsed column in
the row to `missing`, but that doesn't matter or is needed for `(USER |
MISSING)` columns.