Support different dateformats per column
This has been long requested, but the internal structure always assumed
that we'd only ever need a single `Parsers.Options` for an entire file.
For certain parsing features, like date formats, they can vary column to
column, so it makes sense to allow specifying options for a specific
column. This isn't too onerous internally, because we're currently
restricting the option to just dateformats; theoretically we could open
up additional options that may affect a column's type like truestrings,
falsestrings, decimal, and missingstring (though for several of these,
you can already specify multiple options, but this would allow, for
example, specifying a missing string for one column and a different
missingstring for another, without overlap).

I'm limiting to just multiple dateformats for now to kind of kick the
tires and see how minimal of a change we can make this to support the
actual requested feature, but I'm interested to see if we want any other
parsing options to be configurable per column in the future.