Fix AM/PM parsing
Fixes #597.

The issue here was two-fold: one in Parsers.jl for supporting AM/PM
parsing on byte buffers (fix
[here](https://github.com/JuliaData/Parsers.jl/commit/c0ec6ab5815f6c5accac0fae4fc11c47918763fd)),
and an issue here in CSV.jl which involved `CSV.timetype`. This function
takes a user-provided `DateFormat` and tries to figure out if it's meant
to parse `Time`, `Date`, or `DateTime`. The logic, however, was pretty
basic and only checked if the `DateFormat` had a year part for `Date`,
and `H` (hour) part for `Time`, and if a `DateFormat` had both, it was
`DateTime`. The issue is when a user passes `'I'` for 12-based hour
parts, we didn't detect that the date format had a time part. The logic
is now more robust by checking any year/month/day parts for `Date`, and
any hour, minute, second, millisecond parts for `Time`.