Fix missing buffer length check in writecell for Integer (#610)
Should fix #603.

Every once in a while, a bug comes along that just makes you feel
bad. You dig in, you discover the root cause, and you're like, "dammit
past self, what were you thinking?". This is one of those bugs.

A little background: last time I overhauled `CSV.write`, the aim was to
make the code much simpler and more performant. Before, we were doing
individual writes to the output `IO`, which can be really slow depending
on the OS, filesystem, etc. Having a buffer, in our case 4MB worth, can
increase performance because then the bulk of operations are just
storing bytes in an array, with the occasional "flush" out to the `IO`.

In write.jl, we have a handy little `@check` macro that allows a writing
function to say `@check 3` to ensure there are at least 3 bytes of space
left in the buffer to write to. This, again, is an optimization vs.
doing an "end-of-buffer" length check for each byte written. We check
once upfront, then `@inbounds buf[pos] = byte` for as many bytes as we
checked for.

So obviously you're going to get in trouble if you ever forget to call
`@check n` before writing to the buffer, which is what we did here. I'm
sure at the time, my innocent, yet prideful self thought, "ha, this is
just one measly `'-'` byte I'm writing to the buffer, it's probably fine
to avoid the check just this once". But therein lies the problem: a
previous write operation may have, with just the right
alignment/operations, ended writing at the exact end of the 4,194,304
byte buffer. This means the current writing position would be one byte
past the end of the buffer. Then along comes a negative Integer and
pow!, not only do we write past the end of the buffer with `@inbounds
buf[pos] = UInt8('-')`, but we then increment `pos += 1`, grab the
length of the digits in our Integer, call `@check n`, which realizes
we're going to be (are) past the end of the buffer, so it goes to flush
and whoops: we just tried to flush `pos - 1`, which is also one byte
past the end of the buffer.

Goodness. The moral of the story is: always check your buffer lengths,
folks.