Provide a row writing iterator (#611)
* Provide a row writing iterator

Implements suggestion in #608.

This introduces a new structure `CSV.RowWriter`, which is the writing
complement to `CSV.Rows`. It allows taking any Tables.jl-compatible
input, with the same options as `CSV.write`, but instead of writing out
to a file, returns an iterator, where elements are a full csv-written
row.

We mostly had all the machinery in place to provide this, so it's not a
ton of new code. I wondered a bit if the first iteration should produce
column names or just data and if that should be configurable, but
decided to just do column names on first iteration and data thereafter.
I think it's easy enough for callers to drop the first iteration if
not needed anyway, but I'd love to hear people's feedback on that.

I haven't done any tests or docs for this yet; mainly put it up in order
for people to review, so let me know what you think of the design and if
we're missing anything major here.

* lots of tests

* Add docs and error when we overflow teh row buffer

* couple fixes