Change promotion strategy
Instead of keeping track of the position/length of every single cell, we
move to a "reparse" strategy. The problem with tracking poslengths was
the prohibitive memory cost, usually doubling the memory requirements in
the type detection case. This change tracks the starting position of a
chunk and will go back and reparse a single column as string if needed.
I'm not exactly sure how much test coverage we have on this, but I'll
dive in to review and make sure we have some good threaded/non-threaded
cases where promotion is needed.