Add some precompiles
Fixes #464 (or at least improves it quite a bit).

A new precompile.jl file is a script I ran to get some precompile
statements for CSV.jl, Parsers.jl, and SentinelArrays.jl (which seem to
be the biggest targets and ones that live in JuliaData). The Parsers.jl
output ended up being insignificant for now, so that wasn't committed,
but SentinelArrays.jl was. I've added two new "precompile.csv" and
"precompile_small.csv" files that are used for snooping; they include a
column of each type, which should hopefully cover a good chunk of
codepaths we're compiling. We can ajust more later if there are certain
paths that could use it and are causing people problems.

All in all, this cuts TTFP (time-to-first-parse) in half on my machine.