When column names passed manually, ensure we respect starting position
Fixes #668. The issue here is that when column names were passed
manually, the code path that "skipped" to the datarow passed in the
starting position as 1 instead of `pos` variable. This used to not be an
issue because the `pos` was almost always 1 anyway. With `IOBuffer`, we
now start `pos` at `io.ptr`, so we'll have more cases where it's
critical to start reading at right position.