Improve accuracy of estimated rows for multithreaded parsing
I've wanted to do this for a while; previously we were only using the
estimate from the first 10 rows. This hooks into the "chunking" code,
which looks at `tasks` # of chunks of a file to find the start of rows
for each; we now keep track of the # of bytes we saw when doing those
row checks and use those totals plus the original 10 rows to form a
better estimate of the total # of rows.