Ensure dropped columns are ignored in later file processing
Fixes #679; alternative fix to #681. When a column is dropped, we
essentially turn it into a `Missing` column type and ignore it when
parsing. There was a check later in file parsing, however, that said if
no missing values were found in a column, to ensure its type is
`Vector{T}` instead of `Vector{Union{Missing, T}}`. The core problem in
issue #679 was that these dropped columns, while completely `missing`,
didn't get "flagged" as having `missing` values.