Sample column types when finding multithreaded row chunks
This is another optimization I've wanted to do for a while; we're
already checking up to 5 rows from various locations in the file to find
row boundaries for multithreaded parsing, so this just also samples the
types from those locations. The end result is that we already go into
full parsing having checked `tasks * 5` # of rows for column types. The
2 main advantages of adding this are that we can avoid running `detect`
in all tasks in some cases, i.e. if we've already sampled that a column
is `Int64`, then each task will immediately start parsing `Int64`s
instead of having to `detect`, then `allocate`, etc. The other advantage
is if we happen to catch a rogue string value in a part of the file, it
can be promoted immediately instead of having to parse entire chunks as,
say, `Int64`, and then re-parsed later. It's not super robust in that
case because we're only sampling 5 rows from each chunk, but it's better
than nothing and doesn't add extra cost to what we're already doing.

This PR also adds an optimization when a `PooledString` column is
promoted to `String`, we don't need to reparse, but can just allocate
the `String` array and put the ref values in directly according to the
codes.