Fix manually specified categorical type
The issue here is that the user may provide a column type like
`Union{Missing, CategoricalValue}`, but then there may not be any
`missing` values parsed. The issue was the code here assumed that if
`anymissing` flag is set, then there was a `missing` value actually
parsed, but if the user manually specifies a missing type, then we mark
it as `anymissing`. Fixes #706