Don't attempt multithreaded parsing when tasks == 1
The issue here is that a user can pass `threaded=true`, which means
we'll do parsing across `tasks` chunks of a file in parallel, which
causes this error when the number of tasks happens to be 1. This could
be the user passing `tasks=1` manually, in which case they should not
do, or it defaulted to `Threads.nthreads()` and the user's Julia session
started with only 1 thread.

We already do a few other hard-checks for multithreaded parsing, like
ensuring your Julia version is > 1.3, so the fix is pretty simple: if a
user passes `threaded=true`, then we'll also require the `tasks`
parameter to be `> 1`. Fixes #705