Add lines_to_check to allow configuring multithreaded parsing
Fixes #721. When multithreaded parsing, we have to split up a file and
try to find the start of valid rows. By default, we've just been
checking 5 rows to ensure we found the start of a valid row; this PR
allows passing in a custom # of `lines_to_check` in case of problematic
files, usually ones with very large quoted text fields.