Fallback to singled-threaded parsing if file chunking fails
Provides a better experience for cases like #727. On current master, if
there's a mismatch between the number of columns detected in the header
vs. in the rest of the file, the multithreaded chunking operation fails
because it can't accurately detect the right # of columns to be
expected. Also on current master, in this scenario, we throw an error,
which isn't a great user experience since they usually didn't ask
explicitly for threaded parsing anyway. The change in this PR is that if
we fail to determine accurate chunks for multithreaded parsing, then
we'll just fallback to parsing the entire file single-threaded, which
allows full parsing with no errors, though warnings may be emitted.