Return ChainedVector of SentinelArrays as default columns
Instead of SentinelArray of ChainedVector. The main benefit this gives
is if people iterate directly over a column, we'll get the optimized
iteration of ChainedVector directly instead of iterating the
SentinelArray, which then iterates the ChainedVector, which loses some
of the optimization. Shouldn't be a very big functional change
otherwise.