Remove CategoricalArrays/DataFrames dependencies
These have been deprecated for a while now, so in preparation for the
0.8 release, we remove support for the `categorical` keyword argument,
producing `CategoricalArray` columns when the column type is given as
`CategoricalValue`, and `CSV.read` without an explicit sink argument.

I'd like to test a bit more what happens if you provide
CategoricalArray to `CSV.write`, or `CategoricalValue` as a column type.