Avoid ambiguity between user-provided UInt64 column type and PosLen
Fixes #785. The issue here is that we were overloading `allocate(T,
len)` for various types, including our custom `PosLen` type when strings
are just parsed as a position and length. The problem is `PosLen` is
defined as `const PosLen = UInt64`, which means for a user-provided
`UInt64` type, it was allocating a plain `Vector{UInt64}` instead of the
expected `SentinelVector{UInt64}`. We were already special-casing the
allocation of the `PosLen` column, so the fix is pretty easy; create a
new `allocateposlen` instead of relying on the `allocate` overload.