Avoid ambiguity between user-provided UInt64 column type and PosLen
Fixes #785 and alternative solution to #786. As opposed to the solution
in that PR, which caused other issues, in this PR, we change from
defining `const PosLen = UInt64` to `primitive type PosLen 64 end` and
define a few helper conversions. This makes `PosLen` its own proper
type, and avoids the dispatch ambiguity of the original issue.