Ensure we check for commented rows when skipping rows for header/data
Improves #788. In the original issue, a quote character on a commented
row messes the parsing positioning up because it's looking for a closing
quote character. By checking for and skipping commented rows, no matter
the characters present, we ensure parsing integrity. One ramification of
this, however, is that commented rows now "no longer count" when
considering row numbers, i.e. when specifying the `header=2` or
`datarow=4` keyword arguments, because the commented rows are literally
ignored when parsing. This seems fine to me, but probably warrants some
documentation so it's clear.