Don't consider single space as delimiter in auto-detect fallback (#801)
Resolves #799. As noted in that issue, the problem for the OP is in the
auto delimiter detection code, if a delimiter can't be detected that has
an equal # of instances on the rows parsed, it considers the header row
and the most common delimiter detected there. The change in this PR is
to NOT consider a single space `' '` as a valid delimiter in that case
since it's far too common for column names to have spaces, so it's just
throwing off the attempt at detecting a delimiter.