Scrub CSV.File types for Tables.Schema (#811)
Fixes #808. The issue here is that we use an internal `PooledString`
type while parsing to signal a column is currently being pooled. Once
we're done parsing, however, there's no value in keeping `PooledString`
around, and indeed, `Tables.rowtable` even gets confused because it's
expecting a `PooledString` object but we always return `String` objects
when indexing string columns, pooled or not. The fix here is to scrub
these PooledString type columns to correct the `Tables.Schema` on
`CSV.File`. Let's see if CI points out any problems with this approach.