Add automatic support for reading gzipped csv files
Also fixes #733 by using a performant method for processing gzipped
files. The motivation here is to make this more "automatic" for users so
they don't have to fumble around and question whether they should use
GZip.jl or CodecZlib.jl or Zlib.jl or whatever. We know which is the
fastest (CodecZlib.jl), and it's a mature, properly binary-built package
at this point that works cross platform, so I submit that it's a
non-risky dependency to take on for the benefit it provides.

In terms of implementation, we take the same approach as data.table
`fread` here by decompressing the gzip file to a temp file, then
mmapping it. We track the temp files generated in this way and clean
them up at the end of `CSV.File` (I need to check `CSV.Rows` as well),
and we can also rely on having the default of `cleanup=true` for
`tempname()`, which means the julia process itself will try and delete
all its generated temp files when it exits.

Alternative approaches involving reading chunks of a
`GzipDecompressorStream` or having a new `BufferedIOVector <:
AbstractVector{UInt8}` which would buffer reads from a
`GzipDecompressorStream` input were both considered, but at the end of
the day, CSV.jl is hard-coded in several places to work on the input as
a `Vector{UInt8}` explicitly. Parsers.jl is more flexible, but it would
take non-trivial work to allow CSV.jl the same flexibility. Another
factor to consider here is that if we allowed many different `buf` types
(other than just `Vector{UInt8}`), then we dig up the recompilation
problems all over again (that we literally just finished burying with
the removal of type parameters from `Parsers.Options`). That's probably
the biggest factor for me personally to take the `fread` approach of
using tempfiles, in addition to being very performant.

So follow ups before merging are:
  * Check that `CSV.Rows` can clean up tempfiles
  * We can be more strict in a few places of requiring `Vector{UInt8}`
  in CSV.jl
  * Update doc examples to not need `transcode` explicitly