Allow turning off missing value parsing (#865)
* Allow turning off missing value parsing

Fixes #768. We can potentially allow passing `missingstring` per column,
but that seems like a more rare need than turning off missing value
parsing all together, which allows string columns to parse empty strings
correctly by default. The reason per column isn't quite as compelling is
that you can already pass a vector of missing string values to
`missingstring` to check any of the values for `missing`. So it would
only be the use-case of wanting to parse a missing string in one column
_but not_ another where per-column missing strings would be useful.

* add a test