Add extra validation on types/dateformat/pool keyword args (#870)
Fixes #802. The original issue was the user didn't realize the column
name passed in the `types` `Dict` argument didn't quite match the column
name parsed in the file, and `CSV.File` just silently ignored the
"unused" value in the dict and auto-type-detected the column type. This
PR adds an extra validation step once we've parsed the column names to
check that there aren't column names in the `types`/`dateformat`/`pool`
keyword arguments that don't exist in the data itself. If so, an
`ArgumentError` will be thrown, which should help users in this scenario
in the future to realize something is amiss.