Respect typemap in type promotion cases while parsing (#872)
Fixes #871. The issue here is that in our parsevalue! promotion logic,
we weren't considering whether the use may want to typemap a _promoted_
type to another type; in the original issue, a column started out as
detected as integer, then promoted to float, but the user wanted to
typemap any floats to strings. This fix just accounts for any
typemapping that may occur during promotion.