Improve time-to-first-read via precompile + despecialization (#875)
Been wanting/needing to do this for a while. Worked through the
excellent tutorials in
https://timholy.github.io/SnoopCompile.jl/stable/snoopi_deep_parcel/ to
help figure out cases where we were overspecializing certain methods
(`CSV.Context`) and generating useful precompile statements. On my
machine, this reduces the time-to-first-read (TTFR) from ~8s to ~1.5,
while package loading time goes from 3s to 10s, which both seem like
good tradeoffs. To avoid overspecialization of `CSV.Context` method,
which has a huge method body, we introduce a simple `Arg` wrapper struct
to kill inference and make the method signature concrete. The `@refargs`
macro conveniently wraps each arg in a _call_ with `Arg(x)`, and for a
_definition_, it annotates each arg as `x::Arg` and inserts an unwrap as
the first thing in the method body like `x = x[]::Int` where `x` was
declared as `x::Int` in the argument list.