Split Reading and Writing onto own pages
This allows us to have the explanations
and API references on one page, and then
that allows us to have an Examples page
with only examples (no API docs).
We can then bump up the headers so all the
examples appear in the side-bar, to be
easier to find.

Also move the contents list on the homepage
to below the welcome text, and restrict the
contents list on the examples page to only
list the examples (not other pages).

Overall this aims to avoid having deeply-nested
headers (which then don't appear in the sidebar,
making them harder to find), but without
breaking up the docs over too many pages.