Fix a few allocation/inference issues
This was from an effort to play around with precompilation some more.
There's a tricky balance where if we precompile too much of `CSV.File`,
then we run into the Base bug where there are Base methods missing and
cause painful allocations via invoke
(https://github.com/JuliaLang/julia/issues/35972). The state of
precompilation for CSV.jl is thus pretty disappointing: precompiling
currently takes ~7.5s (which is fine, this is the number we're fine to
have keep growing), loading CSV.jl precompiled takes ~2.5s (seems a
little high, but not terrible), but the first call to `CSV.File` takes
~6.5s :sad_face:. At this point, I've spent too much time and don't
understand the ins and outs of precompilation interactions with later
runtime code to debug further, but hopefully we can recruit help from
some experts to take CSV.jl precompilation to the next level. As-is,
it's workable, so let's go with this for now and get this darn release
out.