Refactor internal pooling mechanics (#898)
Fixes #888. One of the core issues is that we weren't "remembering",
when a column's type got detected, whether its pool value was a generic
pool value (like `pool=true`), or provided per-column. This is unique to
pooling, because we don't/can't really apply the pooling until we know
the column type, so it's usually post-context.

The proposal here adds another field `columnspecificpool`, which will
remember this property. The other big issue is that the pooling
mechanics have shifted around a bit throughout the recent internals
refactoring, so we were a little inconsistent in how we were handling
things overall.